+++ app/src/main/java/com/example/myapplication/TempPlayer.java
@@ -22,7 +22,7 @@ public class TempPlayer extends GameObject {
 
     @Override
     public void render(Renderer renderer) {
-        float[] position = getPosition();
+        double[] position = getPosition();
         renderer.drawFilledCircle(position[0], position[1], Color.YELLOW, 10f);
     }
 }
